test payment

+++ payment/views.py
@@ -11,7 +11,7 @@ stripe.api_key = settings.STRIPE_SECRET_KEY
 
 class StripeCheckoutView(APIView):
     def post(self, request):
-        total_amount = request.data.get("total_amt", None) * 100
+        total_amount = int(request.data.get("total_amt", None) * 100)
         payment_method = request.data.get("payment_method", None)
         if total_amount:
             # try:
